+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - add type for every event type
 - use `deltachat-rpc-server.exe` as executable name for the RPC process
 - remove Account.CreateChat()
+- add Bot.IsRunning()
+- add BotRunningErr and RpcRunningErr
 
 ## v0.13.0